+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 0.1.3
 
+- New examples/area_gallery.rb
 - Support `figure(x = nil, y = nil) {}` (`draw_path_new_figure`)
 - Support `closed true` property inside nested figure (`draw_path_close_figure`)
+- Support `line`
 
 ## 0.1.2